+++ admin/src/scenes/volontaires/utils/index.js
@@ -541,3 +541,108 @@ export async function transformVolontairesSchool(data) {
     };
   });
 }
+
+export async function transformInscription(data) {
+  let all = data;
+  const schoolsId = [...new Set(data.map((item) => item.schoolId).filter((e) => e))];
+  if (schoolsId?.length) {
+    const { responses } = await api.esQuery("schoolramses", {
+      query: { bool: { must: { ids: { values: schoolsId } } } },
+      size: ES_NO_LIMIT,
+    });
+    if (responses.length) {
+      const schools = responses[0]?.hits?.hits.map((e) => ({ _id: e._id, ...e._source }));
+      all = data.map((item) => ({ ...item, esSchool: schools?.find((e) => e._id === item.schoolId) }));
+    }
+  }
+  return all.map((data) => {
+    return {
+      _id: data._id,
+      Cohorte: data.cohort,
+      Prénom: data.firstName,
+
+      Nom: data.lastName,
+      "Date de naissance": formatDateFRTimezoneUTC(data.birthdateAt),
+      "Pays de naissance": data.birthCountry || "France",
+      "Ville de naissance": data.birthCity,
+      "Code postal de naissance": data.birthCityZip,
+      "Date de fin de validité de la pièce d'identité": formatDateFRTimezoneUTC(data?.latestCNIFileExpirationDate),
+      Sexe: translate(data.gender),
+      Email: data.email,
+      Téléphone: formatPhoneE164(data.phone, data.phoneZone),
+      "Adresse postale": data.address,
+      "Code postal": data.zip,
+      Ville: data.city,
+      Département: data.department,
+      Région: data.region,
+      Académie: data.academy,
+      Pays: data.country,
+      "Nom de l'hébergeur": data.hostLastName,
+      "Prénom de l'hébergeur": data.hostFirstName,
+      "Lien avec l'hébergeur": data.hostRelationship,
+      "Adresse - étranger": data.foreignAddress,
+      "Code postal - étranger": data.foreignZip,
+      "Ville - étranger": data.foreignCity,
+      "Pays - étranger": data.foreignCountry,
+      Situation: translate(data.situation),
+      Niveau: data.grade,
+      "Type d'établissement": translate(data.esSchool?.type || data.schoolType),
+      "Nom de l'établissement": data.esSchool?.fullName || data.schoolName,
+      "Code postal de l'établissement": data.esSchool?.postcode || data.schoolZip,
+      "Ville de l'établissement": data.esSchool?.city || data.schoolCity,
+      "Département de l'établissement": departmentLookUp[data.esSchool?.department] || data.schoolDepartment,
+      "UAI de l'établissement": data.esSchool?.uai,
+      "Quartier Prioritaire de la ville": translate(data.qpv),
+      "Zone Rurale": translate(isInRuralArea(data)),
+      Handicap: translate(data.handicap),
+      "Bénéficiaire d'un PPS": translate(data.ppsBeneficiary),
+      "Bénéficiaire d'un PAI": translate(data.paiBeneficiary),
+      "Structure médico-sociale": translate(data.medicosocialStructure),
+      "Nom de la structure médico-sociale": data.medicosocialStructureName,
+      "Adresse de la structure médico-sociale": data.medicosocialStructureAddress,
+      "Code postal de la structure médico-sociale": data.medicosocialStructureZip,
+      "Ville de la structure médico-sociale": data.medicosocialStructureCity,
+      "Aménagement spécifique": translate(data.specificAmenagment),
+      "Nature de l'aménagement spécifique": data.specificAmenagmentType,
+      "Aménagement pour mobilité réduite": translate(data.reducedMobilityAccess),
+      "Besoin d'être affecté(e) dans le département de résidence": translate(data.handicapInSameDepartment),
+      "Allergies ou intolérances alimentaires": translate(data.allergies),
+      "Activité de haut-niveau": translate(data.highSkilledActivity),
+      "Nature de l'activité de haut-niveau": data.highSkilledActivityType,
+      "Activités de haut niveau nécessitant d'être affecté dans le département de résidence": translate(data.highSkilledActivityInSameDepartment),
+      "Document activité de haut-niveau ": data.highSkilledActivityProofFiles,
+      "Consentement des représentants légaux": translate(data.parentConsentment),
+      "Droit à l'image": translate(data.imageRight),
+      "Autotest PCR": translate(data.autoTestPCR),
+      "Règlement intérieur": translate(data.rulesYoung),
+      "Fiche sanitaire réceptionnée": translate(data.cohesionStayMedicalFileReceived) || "Non Renseigné",
+      "Statut représentant légal 1": translate(data.parent1Status),
+      "Prénom représentant légal 1": data.parent1FirstName,
+      "Nom représentant légal 1": data.parent1LastName,
+      "Email représentant légal 1": data.parent1Email,
+      "Téléphone représentant légal 1": formatPhoneE164(data.parent1Phone, data.parent1PhoneZone),
+      "Adresse représentant légal 1": data.parent1Address,
+      "Code postal représentant légal 1": data.parent1Zip,
+      "Ville représentant légal 1": data.parent1City,
+      "Département représentant légal 1": data.parent1Department,
+      "Région représentant légal 1": data.parent1Region,
+      "Statut représentant légal 2": translate(data.parent2Status),
+      "Prénom représentant légal 2": data.parent2FirstName,
+      "Nom représentant légal 2": data.parent2LastName,
+      "Email représentant légal 2": data.parent2Email,
+      "Téléphone représentant légal 2": formatPhoneE164(data.parent2Phone, data.parent2PhoneZone),
+      "Adresse représentant légal 2": data.parent2Address,
+      "Code postal représentant légal 2": data.parent2Zip,
+      "Ville représentant légal 2": data.parent2City,
+      "Département représentant légal 2": data.parent2Department,
+      "Région représentant légal 2": data.parent2Region,
+      Motivation: data.motivations,
+      Phase: translate(data.phase),
+      "Créé lé": formatLongDateFR(data.createdAt),
+      "Mis à jour le": formatLongDateFR(data.updatedAt),
+      "Dernière connexion le": formatLongDateFR(data.lastLoginAt),
+      "Statut général": translate(data.status),
+      "Dernier statut le": formatLongDateFR(data.lastStatusAt),
+    };
+  });
+}