+++ api/src/controllers/elasticsearch/schoolramses.js
@@ -1,8 +1,7 @@
 const passport = require("passport");
 const express = require("express");
 const router = express.Router();
-const { ROLES, canSearchInElasticSearch } = require("snu-lib/roles");
-const { ES_NO_LIMIT } = require("snu-lib");
+const { ROLES, ES_NO_LIMIT, canSearchInElasticSearch } = require("snu-lib");
 const { capture } = require("../../sentry");
 const esClient = require("../../es");
 const { ERRORS } = require("../../utils");