+++ app/.eslintrc.js
@@ -11,9 +11,10 @@ module.exports = {
     ecmaVersion: "latest",
     sourceType: "module",
   },
-  plugins: ["react"],
+  plugins: ["react", "prettier"],
   rules: {
     "react/prop-types": 0,
+    "prettier/prettier": ["warn"],
   },
   settings: {
     react: {