+++ app/src/utils/index.js
@@ -51,6 +51,7 @@ export function permissionChangeCohort(y, date) {
 
 export function permissionPhase1(y) {
   if (!permissionApp(y)) return false;
+  if (y.cohort === "à venir") return false;
   return (
     (![YOUNG_STATUS.WAITING_VALIDATION, YOUNG_STATUS.WAITING_CORRECTION, YOUNG_STATUS.WAITING_LIST, YOUNG_STATUS.WITHDRAWN].includes(y.status) &&
       y.statusPhase1 !== YOUNG_STATUS_PHASE1.NOT_DONE &&